added unit test
